feat: Add cloud video URL field
Introduces a new field `cloudVideoUrl` to the song data model. This allows for storing direct download links to high-quality video files from cloud storage services.

The field is integrated into the Add Song form and the Song Detail editing view. It is also made available in the Streaming page component.